requirements.txt
aws-cdk-lib==2.110.0
constructs>=10.0.0,<11.0.0
python-dotenv==1.0.1
moto-ext
responses
requests-mock
lambda-local
python-lambda-local
botocore==1.34.144
boto3==1.34.144
svix-ksuid